move all core sequencer code to lib/sequencer
