+++ src/www/lib/sequencer/bpm-utilities.js
