+++ src/www/lib/sequencer/worker-setInterval.js
